Minor Tweaks
Just a few small things. Biggest change is the slight buff to the log
helper.
