+++ src/main/java/com/infinityraider/infinitylib/block/BlockTileCustomRenderedBase.java
@@ -7,7 +7,6 @@ import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
-import net.minecraft.world.World;
 
 @SuppressWarnings("unused")
 public abstract class BlockTileCustomRenderedBase<T extends TileEntityBase> extends BlockBaseTile<T> implements ICustomRenderedBlockWithTile<T> {
